LOL, demo using a different, exotic language thru custom webpack
