--- packages/template-typescript-minimal/app/webpack.config.js
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2021, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-const config = require('pwa-kit-build/configs/webpack/config')
-
-module.exports = config
