+++ packages/template-typescript-minimal/package.json
@@ -8,6 +8,8 @@
     "private": true,
     "devDependencies": {
         "@loadable/component": "^5.15.0",
+        "coffee-loader": "^3.0.0",
+        "coffeescript": "^2.6.1",
         "cross-env": "^5.2.0",
         "cross-fetch": "^3.1.4",
         "pwa-kit-build": "^2.0.0-dev.3",
